requirements.txt
-i https://pypi.org/simple
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
chardet==3.0.4
click==7.0
cycler==0.10.0
facenet==1.0.3
flask==1.0.2
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
keras-applications==1.0.7
keras-preprocessing==1.0.9
kiwisolver==1.1.0
markdown==3.1
markupsafe==1.1.1
matplotlib==3.0.3
mock==2.0.0
numpy==1.16.1
opencv-python==4.1.0.25
pbr==5.2.0
pillow==6.0.0
pipfile==0.0.2
protobuf==3.7.1
psutil==5.6.2
pyparsing==2.4.0
python-dateutil==2.8.0
requests==2.21.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow==1.13.1
termcolor==1.1.0
toml==0.10.0
urllib3==1.24.2
werkzeug==0.15.2
wheel==0.33.1 ; python_version >= '3'
